K Bandit Problem
An implementation of the K-bandit reinforcement
learning problem.

We use the sample average method to estimate
action values. The optimal values q*(a) are
drawn from a standard normal distrubtion and
then the actual rewards are selected from
normal distributions with mean q*(a) and
variance one.

The eta value is varied to randomly explore
different actions. Ties are broken by picking
an action uniformly at random