K Bandit Non-stationary problem
Perform a random walk over the q*(a) optimal
values. Use an expoential recency-weighted
action-value scheme to decay older states
and weight them using a constant value.

This means that convergence is not guaranteed
as the sum of the squaured values of the sequence
of these constants is not less than infinity but,
for real-world non-stationary problems, this is
actually desired.